Move sub-components of Game into GameSheet directory
