+++ src/components/GameSheet/Check/Check.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import Button from '../../UI/atoms/Button/Button';
+import Button from '../../../UI/atoms/Button/Button';
 import '../Answers/Answers.css';
-import withAnimation from '../../HOCs/withAnimation/withAnimation';
-import { useStatus } from '../../context/StatusContext';
-import { useAnswer } from '../../context/AnswerContext';
-import { useSum } from '../../context/SumContext';
-import { useScore } from '../../context/ScoreContext';
+import withAnimation from '../../../HOCs/withAnimation/withAnimation';
+import { useStatus } from '../../../context/StatusContext';
+import { useAnswer } from '../../../context/AnswerContext';
+import { useSum } from '../../../context/SumContext';
+import { useScore } from '../../../context/ScoreContext';
 
 const Check: React.FC = (): JSX.Element => {
   const { setGameStatus } = useStatus();
