+++ src/components/GameSheet/Result/Result.scss
