+++ src/components/GameSheet/Score/Score.scss
