+++ src/components/GameSheet/Sum/Sum.scss
